requirements.txt
# Triple-Stream Forensic Framework — dependencies
# Install: pip install -r requirements.txt

# Deep learning
torch>=1.12.0
torchvision>=0.13.0

# Forensic feature extraction
opencv-python>=4.6.0
Pillow>=9.0.0
scipy>=1.9.0
numpy>=1.22.0

# Retrieval indexing
faiss-cpu>=1.7.3          # use faiss-gpu if CUDA is available

# Interpretability
grad-cam>=1.4.6

# Evaluation / visualisation
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.4.0

# Progress bars
tqdm>=4.64.0
